requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
ebooklib
pytesseract
Pillow
beautifulsoup4
jinja2